[4044] Fix avatar in navigation drawer (#4050)
* [4044] Fix avatar in navigation drawer

* Update change log
